Converted descriptor.sets to idiomatic Kotlin. - Fixed bug in LoadingEffectToAddDefinition, where it was choosing an   arbitrary bundle for a method instead of the one that was present in   the original action.  If it chose a bundle that was added as an alias   within the same function that added the definition, then on replay it   would create a method for that bundle and add the definition, prior to   noting that it was an alias of the other bundle, thereby failing to   load the module. - Converted set/map bin descriptors' myLevel field from Byte->Int.  Also   the shift field. - Globally corrected space-tab -> tab in Java/Kotlin/Avail code. - Upgraded kotlin_version to 1.3.72 in gradle.properties.  Also added   kotlin.incremental=true - L1Decompiler now treats literal tuples as list phrases if needed. - In Set Tests "each of satisfies" and "each of satisfies (macro)",   removed dependence on set iteration order. - Added simple SetTest.java.

+++ src/main/kotlin/com/avail/descriptor/representation/A_BasicObject.kt
@@ -228,7 +228,7 @@ interface A_BasicObject : JSONFriendly {
 	fun setBinAddingElementHashLevelCanDestroy(
 		elementObject: A_BasicObject,
 		elementObjectHash: Int,
-		myLevel: Byte,
+		myLevel: Int,
 		canDestroy: Boolean): A_BasicObject
 
 	/**
@@ -254,7 +254,7 @@ interface A_BasicObject : JSONFriendly {
 	fun binRemoveElementHashLevelCanDestroy(
 		elementObject: A_BasicObject,
 		elementObjectHash: Int,
-		myLevel: Byte,
+		myLevel: Int,
 		canDestroy: Boolean): AvailObject
 
 	/**
